// NetscriptDefinitions.d.ts
interface Player {
    hacking: number;
    hp: number;
    max_hp: number;
    strength: number;
    defense: number;
    dexterity: number;
    agility: number;
    charisma: number;
    intelligence: number;
    hacking_chance_mult: number;
    hacking_speed_mult: number;
    hacking_money_mult: number;
    hacking_grow_mult: number;
    hacking_exp: number;
    strength_exp: number;
    defense_exp: number;
    dexterity_exp: number;
    agility_exp: number;
    charisma_exp: number;
    hacking_mult: number;
    strength_mult: number;
    defense_mult: number;
    dexterity_mult: number;
    agility_mult: number;
    charisma_mult: number;
    hacking_exp_mult: number;
    strength_exp_mult: number;
    defense_exp_mult: number;
    dexterity_exp_mult: number;
    agility_exp_mult: number;
    charisma_exp_mult: number;
    company_rep_mult: number;
    faction_rep_mult: number;
    numPeopleKilled: number;
    money: number;
    city: string;
    location: string;
    companyName: string;
    crime_money_mult: number;
    crime_success_mult: number;
    isWorking: boolean;
    workType: string;
    currentWorkFactionName: string;
    currentWorkFactionDescription: string;
    workHackExpGainRate: number;
    workStrExpGainRate: number;
    workDefExpGainRate: number;
    workDexExpGainRate: number;
    workAgiExpGainRate: number;
    workChaExpGainRate: number;
    workRepGainRate: number;
    workMoneyGainRate: number;
    workMoneyLossRate: number;
    workHackExpGained: number;
    workStrExpGained: number;
    workDefExpGained: number;
    workDexExpGained: number;
    workAgiExpGained: number;
    workChaExpGained: number;
    workRepGained: number;
    workMoneyGained: number;
    createProgramName: string;
    createProgramReqLvl: number;
    className: string;
    crimeType: string;
    work_money_mult: number;
    hacknet_node_money_mult: number;
    hacknet_node_purchase_cost_mult: number;
    hacknet_node_ram_cost_mult: number;
    hacknet_node_core_cost_mult: number;
    hacknet_node_level_cost_mult: number;
    hasWseAccount: boolean;
    hasTixApiAccess: boolean;
    has4SData: boolean;
    has4SDataTixApi: boolean;
    bladeburner_max_stamina_mult: number;
    bladeburner_stamina_gain_mult: number;
    bladeburner_analysis_mult: number;
    bladeburner_success_chance_mult: number;
    bitNodeN: number;
    totalPlaytime: number;
    playtimeSinceLastAug: number;
    playtimeSinceLastBitnode: number;
    jobs: any;
    factions: string[];
    tor: boolean;
    hasCorporation: boolean;
}
// ...
export interface GangGenInfo {
    /** Name of faction that the gang belongs to ("Slum Snakes", etc.) */
    faction: string;
    /** Indicating whether or not it's a hacking gang */
    isHacking: boolean;
    /** Money earned per game cycle */
    moneyGainRate: number;
    /** Gang's power for territory warfare */
    power: number;
    /** Gang's respect */
    respect: number;
    /** Respect earned per game cycle */
    respectGainRate: number;
    /** Amount of territory held */
    territory: number;
    /** Clash chance */
    territoryClashChance: number;
    /** Gang's wanted level */
    wantedLevel: number;
    /** Wanted level gained/lost per game cycle (negative for losses) */
    wantedLevelGainRate: number;
    /** Indicating if territory warfare is enabled */
    territoryWarfareEngaged: boolean;
    /** Number indicating the current wanted penalty */
    wantedPenalty: number;
}
// ...
export interface GangTaskStats {
    /** Task name */
    name: string;
    /** Task Description */
    desc: string;
    /** Is a task of a hacking gang */
    isHacking: boolean;
    /** Is a task of a combat gang */
    isCombat: boolean;
    /** Base respect earned */
    baseRespect: number;
    /** Base wanted earned */
    baseWanted: number;
    /** Base money earned */
    baseMoney: number;
    /** Hacking skill impact on task scaling */
    hackWeight: number;
    /** Strength skill impact on task scaling */
    strWeight: number;
    /** Defense skill impact on task scaling */
    defWeight: number;
    /** Dexterity skill impact on task scaling */
    dexWeight: number;
    /** Agility skill impact on task scaling */
    agiWeight: number;
    /** Charisma skill impact on task scaling */
    chaWeight: number;
    /** Number representing the difficulty of the task */
    difficulty: number;
    /** Territory impact on task scaling */
    territory: GangTerritory;
}
// ...
export interface GangTerritory {
    /** Money gain impact on task scaling */
    money: number;
    /** Respect gain impact on task scaling */
    respect: number;
    /** Wanted gain impact on task scaling */
    wanted: number;
}

/**
 * @public
 */
export interface GangMemberInfo {
    /** Name of the gang member */
    name: string;
    /** Currently assigned task */
    task: string;
    earnedRespect: number;

    /** Hack skill level */
    hack: number;
    /** Strength skill level */
    str: number;
    /** Defense skill level */
    def: number;
    /** Dexterity skill level */
    dex: number;
    /** Agility skill level */
    agi: number;
    /** Charisma skill level */
    cha: number;

    /** Current hack experience */
    hack_exp: number;
    /** Current strength experience */
    str_exp: number;
    /** Current defense experience */
    def_exp: number;
    /** Current dexterity experience */
    dex_exp: number;
    /** Current agility experience */
    agi_exp: number;
    /** Current charisma experience */
    cha_exp: number;

    /** Hack multiplier from equipment */
    hack_mult: number;
    /** Strength multiplier from equipment */
    str_mult: number;
    /** Defense multiplier from equipment */
    def_mult: number;
    /** Dexterity multiplier from equipment */
    dex_mult: number;
    /** Agility multiplier from equipment */
    agi_mult: number;
    /** Charisma multiplier from equipment */
    cha_mult: number;

    /** Hack multiplier from ascensions */
    hack_asc_mult: number;
    /** Strength multiplier from ascensions */
    str_asc_mult: number;
    /** Defense multiplier from ascensions */
    def_asc_mult: number;
    /** Dexterity multiplier from ascensions */
    dex_asc_mult: number;
    /** Agility multiplier from ascensions */
    agi_asc_mult: number;
    /** Charisma multiplier from ascensions */
    cha_asc_mult: number;

    /** Total earned hack experience */
    hack_asc_points: number;
    /** Total earned strength experience */
    str_asc_points: number;
    /** Total earned defense experience */
    def_asc_points: number;
    /** Total earned dexterity experience */
    dex_asc_points: number;
    /** Total earned agility experience */
    agi_asc_points: number;
    /** Total earned charisma experience */
    cha_asc_points: number;

    upgrades: string[];
    augmentations: string[];

    respectGain: number;
    wantedLevelGain: number;
    moneyGain: number;
}